Refactor + Javadoc

+++ src/fr/upem/spacekaira/util/Util.java
@@ -1,7 +1,12 @@
 package fr.upem.spacekaira.util;
 
 public class Util {
-    public static String getTimeLeftAsString(long startTime, long gameDuration) {
+    /**
+     * @param startTime Start time in milliseconds
+     * @param gameDuration Game duration in seconds
+     * @return Example: for 1m5s, return 1:05
+     */
+    public static String getTimeLeftAsString(long startTime, int gameDuration) {
         long currentTime = (long) Math.floor(System.currentTimeMillis() / 1000);
         startTime = (long) Math.floor(startTime / 1000);
 
@@ -12,6 +17,11 @@ public class Util {
         return String.format("%d:%02d", minutes, seconds);
     }
 
+    /**
+     * @param startTime Start time in milliseconds
+     * @param gameDuration Game duration in seconds
+     * @return Example: for 1m10s, returns 70
+     */
     public static long getTimeLeftAsLong(long startTime, int gameDuration) {
         return (gameDuration * 1000 - System.currentTimeMillis() + startTime) / 1000;
     }
